+++ io.openems.edge.common/src/io/openems/edge/common/modbusslave/ModbusRecordUint16Reserved.java
@@ -3,7 +3,7 @@ package io.openems.edge.common.modbusslave;
 public class ModbusRecordUint16Reserved extends ModbusRecordUint16 {
 
 	public ModbusRecordUint16Reserved(int offset) {
-		super(offset, "Reserved", null);
+		super(offset, "Reserved", (Short) null);
 	}
 
 	@Override
